Move BaseEvent to shared folder
