+++ src/core/eventbus/index.ts
@@ -1,3 +1,2 @@
 export * from './event.module.js';
 export * from './services/event.service.js';
-export * from './types/base-event.js';
