+++ src/core/eventbus/types/util.types.ts
@@ -1,4 +1,4 @@
-import { BaseEvent } from './base-event.js';
+import { BaseEvent } from '../../../shared/events/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type Constructor<T = unknown> = new (...args: any[]) => T;
